+++ CHANGELOG.md
@@ -86,6 +86,7 @@ GRDB adheres to [Semantic Versioning](https://semver.org/), with one exception: 
 ## Next Release
 
 - **New**: `DatabaseRegion(table: tableName)` is deprecated. Use `Table(tableName)` instead.
+- **New**: The `ABS` and `LENGTH` SQL functions are now available on association aggregates, through `abs(aggregate)` and `length(aggregate)`.
 
 ## 5.16.0